Limit the DNSSEC cache.
Make the DNSSEC cache a LRU and limit it.

Closes #13
